fix type of data in measurement table

+++ main/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 # Create your models here.
 class Measurement(models.Model):
-    data = models.IntegerField()
+    data = models.DecimalField(max_digits=7, decimal_places=2)
     longitude = models.DecimalField(max_digits=9, decimal_places=6)
     latitude = models.DecimalField(max_digits=8, decimal_places=6)
     date_time = models.DateTimeField()
